Test sign-up form
This required a little hackery because Devise has methods on the
`Devise::RegistrationssController` that otherwise aren't available in
specs.

Since so far we only need these on one template, I've stubbed them in
one place, but when we test `#edit`, we'll want to DRY it up.

See https://stackoverflow.com/questions/14426746/testing-devise-views-with-rspec